remove download codes from timesheet to create a separate component. Then import component to timesheet instead.
